fix: move api logic into common server files functions
