+++ src/app/whiteboards/page.tsx
@@ -1,7 +1,7 @@
-import { getWhiteboards } from "@/lib/whiteboards";
+import { getWhiteboardsFormatted } from "@/lib/whiteboards";
 import { Whiteboards } from "@/components/whiteboards/whiteboards";
 
 export default async function Home() {
-  const whiteboards = await getWhiteboards();
+  const whiteboards = await getWhiteboardsFormatted();
   return <Whiteboards whiteboards={whiteboards ?? []} />;
 }
